+++ src/KlarnaOrderManagement.php
@@ -9,7 +9,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 /**
  * Required minimums and constants
  */
-define( 'KLARNA_ORDER_MANAGEMENT_VERSION', '1.0.3' );
+define( 'KLARNA_ORDER_MANAGEMENT_VERSION', '1.1.0' );
 
 define( 'KLARNA_ORDER_MANAGEMENT_MIN_PHP_VER', '5.3.0' );
 define( 'KLARNA_ORDER_MANAGEMENT_MIN_WC_VER', '3.3.0' );
@@ -22,6 +22,7 @@ use Krokedil\KlarnaOrderManagement\Request\Post\RequestPostCapture;
 use Krokedil\KlarnaOrderManagement\Request\Patch\RequestPatchUpdate;
 use Krokedil\KlarnaOrderManagement\Request\Post\RequestPostCancel;
 use Krokedil\KlarnaOrderManagement\MetaBox;
+use Krokedil\KlarnaOrderManagement\Ajax;
 
 /**
  * Klarna Order Management class.
@@ -43,6 +44,13 @@ class KlarnaOrderManagement {
 	 * @var MetaBox $metabox
 	 */
 	public $metabox;
+
+	/**
+	 * Klarna Order Management AJAX handler.
+	 *
+	 * @var Ajax $ajax
+	 */
+	public $ajax;
 
 	/**
 	 * Klarna Order Management plugin instance.
@@ -102,6 +110,7 @@ class KlarnaOrderManagement {
 
 		$this->settings = new Settings();
 		$this->metabox  = new MetaBox( $this );
+		$this->ajax     = new Ajax();
 
 		// Add refunds support to Klarna Payments or Klarna Checkout gateways. If not one of these plugins, do nothing.
 		switch ( $this->plugin_instance ) {
@@ -150,6 +159,9 @@ class KlarnaOrderManagement {
 	public function enqueue_admin() {
 		wp_enqueue_style( 'kom-admin-style', plugin_dir_url( __FILE__ ) . 'assets/css/klarna-order-management.css', array(), '1.0.0' );
 		wp_enqueue_script( 'kom-admin-js', plugin_dir_url( __FILE__ ) . 'assets/js/klarna-order-management.js', array( 'jquery' ), '1.0.0', true );
+		if ( isset( $this->metabox ) && method_exists( $this->metabox, 'maybe_localize_script' ) ) {
+			$this->metabox->maybe_localize_script( 'kom-admin-js' );
+		}
 	}
 
 	/**