+++ changelog.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ------------------
+## [1.1.0] - 2025-10-20
+
+### Added
+
+* Integrated the metabox package to enhance the order management metabox functionality.
+* Added filter `kom_skip_scheduled_actions` to allow disabling the display of scheduled actions on the admin order page.
+
 ## [1.0.3] - 2025-10-01
 
 ### Fixed